dapp-refactor: reorganized token-create-custom
