+++ system-contract-dapp-playground/src/components/contract-interaction/hts/token-create-custom/methods/AssociateHederaToken/index.tsx
@@ -25,20 +25,20 @@ import { useState, useEffect } from 'react';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
 import { generatedRandomUniqueKey } from '@/utils/common/helpers';
 import { TransactionResult } from '@/types/contract-interactions/HTS';
-import { TRANSACTION_PAGE_SIZE } from '../../shared/states/commonStates';
-import { useToastSuccessful } from '../../../../../hooks/useToastSuccessful';
-import { handleAPIErrors } from '../../../../common/methods/handleAPIErrors';
-import { usePaginatedTxResults } from '../../../../../hooks/usePaginatedTxResults';
-import TokenAddressesInputForm from '../../shared/components/TokenAddressesInputForm';
-import { TransactionResultTable } from '../../../../common/components/TransactionResultTable';
-import { handleSanitizeHederaFormInputs } from '../../../../common/methods/handleSanitizeFormInputs';
+import { TRANSACTION_PAGE_SIZE } from '../../../shared/states/commonStates';
+import { useToastSuccessful } from '../../../../../../hooks/useToastSuccessful';
+import { handleAPIErrors } from '../../../../../common/methods/handleAPIErrors';
+import { usePaginatedTxResults } from '../../../../../../hooks/usePaginatedTxResults';
+import TokenAddressesInputForm from '../../../shared/components/TokenAddressesInputForm';
+import { TransactionResultTable } from '../../../../../common/components/TransactionResultTable';
+import { handleSanitizeHederaFormInputs } from '../../../../../common/methods/handleSanitizeFormInputs';
 import { CONTRACT_NAMES, HEDERA_TRANSACTION_RESULT_STORAGE_KEYS } from '@/utils/common/constants';
-import { SharedFormInputField, SharedExecuteButton } from '../../shared/components/ParamInputForm';
-import { useUpdateTransactionResultsToLocalStorage } from '../../../../../hooks/useUpdateLocalStorage';
+import { SharedFormInputField, SharedExecuteButton } from '../../../shared/components/ParamInputForm';
+import { useUpdateTransactionResultsToLocalStorage } from '../../../../../../hooks/useUpdateLocalStorage';
 import { htsTokenAssociateParamFields } from '@/utils/contract-interactions/HTS/token-create-custom/constant';
 import { associateHederaTokensToAccounts } from '@/api/hedera/hts-interactions/tokenCreateCustom-interactions';
-import useFilterTransactionsByContractAddress from '../../../../../hooks/useFilterTransactionsByContractAddress';
-import { handleRetrievingTransactionResultsFromLocalStorage } from '../../../../common/methods/handleRetrievingTransactionResultsFromLocalStorage';
+import useFilterTransactionsByContractAddress from '../../../../../../hooks/useFilterTransactionsByContractAddress';
+import { handleRetrievingTransactionResultsFromLocalStorage } from '../../../../../common/methods/handleRetrievingTransactionResultsFromLocalStorage';
 
 interface PageProps {
   baseContract: Contract;
