+++ system-contract-dapp-playground/src/components/contract-interaction/hts/token-create-custom/methods/FungibleTokenCreate/index.tsx
@@ -23,23 +23,23 @@ import { Contract } from 'ethers';
 import { useToast } from '@chakra-ui/react';
 import { useState, useEffect } from 'react';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
-import { handleAPIErrors } from '../../../../common/methods/handleAPIErrors';
-import { useToastSuccessful } from '../../../../../hooks/useToastSuccessful';
-import { usePaginatedTxResults } from '../../../../../hooks/usePaginatedTxResults';
-import { SharedSigningKeysComponent } from '../../shared/components/SigningKeysForm';
-import { TransactionResultTable } from '../../../../common/components/TransactionResultTable';
-import { handleSanitizeHederaFormInputs } from '../../../../common/methods/handleSanitizeFormInputs';
+import { handleAPIErrors } from '../../../../../common/methods/handleAPIErrors';
+import { useToastSuccessful } from '../../../../../../hooks/useToastSuccessful';
+import { usePaginatedTxResults } from '../../../../../../hooks/usePaginatedTxResults';
+import { SharedSigningKeysComponent } from '../../../shared/components/SigningKeysForm';
+import { TransactionResultTable } from '../../../../../common/components/TransactionResultTable';
+import { handleSanitizeHederaFormInputs } from '../../../../../common/methods/handleSanitizeFormInputs';
 import { CONTRACT_NAMES, HEDERA_TRANSACTION_RESULT_STORAGE_KEYS } from '@/utils/common/constants';
-import { useUpdateTransactionResultsToLocalStorage } from '../../../../../hooks/useUpdateLocalStorage';
+import { useUpdateTransactionResultsToLocalStorage } from '../../../../../../hooks/useUpdateLocalStorage';
 import { createHederaFungibleToken } from '@/api/hedera/hts-interactions/tokenCreateCustom-interactions';
 import { htsTokenCreateParamFields } from '@/utils/contract-interactions/HTS/token-create-custom/constant';
-import useFilterTransactionsByContractAddress from '../../../../../hooks/useFilterTransactionsByContractAddress';
-import { handleRetrievingTransactionResultsFromLocalStorage } from '../../../../common/methods/handleRetrievingTransactionResultsFromLocalStorage';
+import useFilterTransactionsByContractAddress from '../../../../../../hooks/useFilterTransactionsByContractAddress';
+import { handleRetrievingTransactionResultsFromLocalStorage } from '../../../../../common/methods/handleRetrievingTransactionResultsFromLocalStorage';
 import {
   SharedFormInputField,
   SharedFormButton,
   SharedExecuteButtonWithFee,
-} from '../../shared/components/ParamInputForm';
+} from '../../../shared/components/ParamInputForm';
 import {
   CommonKeyObject,
   TransactionResult,
@@ -49,7 +49,7 @@ import {
   HederaTokenKeyTypes,
   TRANSACTION_PAGE_SIZE,
   HederaTokenKeyValueType,
-} from '../../shared/states/commonStates';
+} from '../../../shared/states/commonStates';
 
 interface PageProps {
   baseContract: Contract;
