+++ system-contract-dapp-playground/src/components/contract-interaction/hts/token-create-custom/methods/MintHederaToken/index.tsx
@@ -25,23 +25,23 @@ import { useEffect, useMemo, useState } from 'react';
 import { CommonErrorToast } from '@/components/toast/CommonToast';
 import { generatedRandomUniqueKey } from '@/utils/common/helpers';
 import { TransactionResult } from '@/types/contract-interactions/HTS';
-import { TRANSACTION_PAGE_SIZE } from '../../shared/states/commonStates';
-import MetadataInputForm from '../../shared/components/MetadataInputForm';
-import { handleAPIErrors } from '../../../../common/methods/handleAPIErrors';
-import { useToastSuccessful } from '../../../../../hooks/useToastSuccessful';
-import { usePaginatedTxResults } from '../../../../../hooks/usePaginatedTxResults';
-import { TransactionResultTable } from '../../../../common/components/TransactionResultTable';
-import { handleSanitizeHederaFormInputs } from '../../../../common/methods/handleSanitizeFormInputs';
+import { TRANSACTION_PAGE_SIZE } from '../../../shared/states/commonStates';
+import MetadataInputForm from '../../../shared/components/MetadataInputForm';
+import { handleAPIErrors } from '../../../../../common/methods/handleAPIErrors';
+import { useToastSuccessful } from '../../../../../../hooks/useToastSuccessful';
+import { usePaginatedTxResults } from '../../../../../../hooks/usePaginatedTxResults';
+import { TransactionResultTable } from '../../../../../common/components/TransactionResultTable';
+import { handleSanitizeHederaFormInputs } from '../../../../../common/methods/handleSanitizeFormInputs';
 import { CONTRACT_NAMES, HEDERA_TRANSACTION_RESULT_STORAGE_KEYS } from '@/utils/common/constants';
-import { useUpdateTransactionResultsToLocalStorage } from '../../../../../hooks/useUpdateLocalStorage';
+import { useUpdateTransactionResultsToLocalStorage } from '../../../../../../hooks/useUpdateLocalStorage';
 import { htsTokenMintParamFields } from '@/utils/contract-interactions/HTS/token-create-custom/constant';
-import useFilterTransactionsByContractAddress from '../../../../../hooks/useFilterTransactionsByContractAddress';
-import { handleRetrievingTransactionResultsFromLocalStorage } from '../../../../common/methods/handleRetrievingTransactionResultsFromLocalStorage';
+import useFilterTransactionsByContractAddress from '../../../../../../hooks/useFilterTransactionsByContractAddress';
+import { handleRetrievingTransactionResultsFromLocalStorage } from '../../../../../common/methods/handleRetrievingTransactionResultsFromLocalStorage';
 import {
   SharedFormButton,
   SharedExecuteButton,
   SharedFormInputField,
-} from '../../shared/components/ParamInputForm';
+} from '../../../shared/components/ParamInputForm';
 import {
   mintHederaToken,
   mintHederaTokenToAddress,
