+++ system-contract-dapp-playground/src/components/contract-interaction/hts/token-create-custom/methods/index.tsx
@@ -37,9 +37,7 @@ const HederaTokenCreateMethods = ({ baseContract, method }: PageProps) => {
       {method === 'grantKYC' && <GrantTokenKYC baseContract={baseContract} />}
       {method === 'tokenAssociation' && <AssociateHederaToken baseContract={baseContract} />}
       {method === 'fungibleTokenCreate' && <FungibleTokenCreate baseContract={baseContract} />}
-      {method === 'non-fungibleTokenCreate' && (
-        <NonFungibleTokenCreate baseContract={baseContract} />
-      )}
+      {method === 'non-fungibleTokenCreate' && <NonFungibleTokenCreate baseContract={baseContract} />}
     </>
   );
 };
